requirements.txt
Flask>=3.0
flask-socketio
python-socketio
python-engineio
numpy
Pillow
opencv-python
reportlab
librosa
soundfile
